+++ crates/biome_css_analyze/src/lint/nursery/no_duplicate_font_names.rs
@@ -2,12 +2,10 @@ use std::collections::HashSet;
 
 use biome_analyze::{context::RuleContext, declare_rule, Ast, Rule, RuleDiagnostic, RuleSource};
 use biome_console::markup;
-use biome_css_syntax::{
-    AnyCssGenericComponentValue, AnyCssValue, CssGenericComponentValueList, CssGenericProperty,
-};
-use biome_rowan::{AstNode, SyntaxNodeCast, TextRange};
+use biome_css_syntax::{AnyCssGenericComponentValue, AnyCssValue, CssGenericProperty};
+use biome_rowan::{AstNode, TextRange};
 
-use crate::utils::{is_css_variable, is_font_family_keyword, is_font_shorthand_keyword};
+use crate::utils::{find_font_family, is_font_family_keyword};
 
 declare_rule! {
     /// Disallow duplicate names within font families.
@@ -151,70 +149,4 @@ impl Rule for NoDuplicateFontNames {
             }),
         )
     }
-}
-
-// Get the font-families within a `font` shorthand property value.
-fn find_font_family(value: CssGenericComponentValueList) -> Vec<AnyCssValue> {
-    let mut font_families: Vec<AnyCssValue> = Vec::new();
-    for v in value {
-        let lower_case_value = v.text().to_lowercase();
-
-        // Ignore CSS variables
-        if is_css_variable(&lower_case_value) {
-            continue;
-        }
-
-        // Ignore keywords for other font parts
-        if is_font_shorthand_keyword(&lower_case_value)
-            && !is_font_family_keyword(&lower_case_value)
-        {
-            continue;
-        }
-
-        // Ignore font-sizes
-        if matches!(
-            v,
-            AnyCssGenericComponentValue::AnyCssValue(AnyCssValue::AnyCssDimension(_))
-        ) {
-            continue;
-        }
-
-        // Ignore anything come after a <font-size>/, because it's a line-height
-        if let Some(prev_node) = v.syntax().prev_sibling() {
-            if let Some(prev_prev_node) = prev_node.prev_sibling() {
-                if let Some(slash) = prev_node.cast::<AnyCssGenericComponentValue>() {
-                    if let Some(size) = prev_prev_node.cast::<AnyCssGenericComponentValue>() {
-                        if matches!(
-                            size,
-                            AnyCssGenericComponentValue::AnyCssValue(AnyCssValue::AnyCssDimension(
-                                _
-                            ))
-                        ) && matches!(slash, AnyCssGenericComponentValue::CssGenericDelimiter(_))
-                        {
-                            continue;
-                        }
-                    }
-                };
-            }
-        }
-
-        // Ignore number values
-        if matches!(
-            v,
-            AnyCssGenericComponentValue::AnyCssValue(AnyCssValue::CssNumber(_))
-        ) {
-            continue;
-        }
-
-        match v {
-            AnyCssGenericComponentValue::CssGenericDelimiter(_) => continue,
-            AnyCssGenericComponentValue::AnyCssValue(css_value) => match css_value {
-                AnyCssValue::CssIdentifier(_) | AnyCssValue::CssString(_) => {
-                    font_families.push(css_value)
-                }
-                _ => continue,
-            },
-        }
-    }
-    font_families
 }
